+++ app/src/main/java/com/infomaniak/drive/ui/menu/OfflineFileFragment.kt
@@ -19,11 +19,8 @@ package com.infomaniak.drive.ui.menu
 
 import android.os.Bundle
 import android.view.View
-import androidx.core.view.isGone
 import com.infomaniak.drive.R
 import com.infomaniak.drive.data.cache.FileController
-import com.infomaniak.drive.data.models.BulkOperationType
-import com.infomaniak.drive.data.models.File
 import com.infomaniak.drive.ui.fileList.multiSelect.MultiSelectActionsBottomSheetDialogArgs
 import com.infomaniak.drive.ui.fileList.multiSelect.OfflineMultiSelectActionsBottomSheetDialog
 import com.infomaniak.drive.utils.Utils
@@ -39,17 +36,12 @@ open class OfflineFileFragment : FileSubTypeListFragment() {
         super.onViewCreated(view, savedInstanceState)
         collapsingToolbarLayout.title = getString(R.string.offlineFileTitle)
         swipeRefreshLayout.isEnabled = false
-        setupMultiSelectLayout()
     }
 
     private fun initParams() {
         downloadFiles = DownloadFiles()
         setNoFilesLayout = SetNoFilesLayout()
         folderId = Utils.OTHER_ROOT_ID
-    }
-
-    private fun setupMultiSelectLayout() {
-        multiSelectLayout?.selectAllButton?.isGone = true
     }
 
     override fun onMenuButtonClicked() {
@@ -63,19 +55,6 @@ open class OfflineFileFragment : FileSubTypeListFragment() {
                 isAllSelected = isAllSelected
             ).toBundle()
         }.show(childFragmentManager, "ActionOfflineMultiSelectBottomSheetDialog")
-    }
-
-    override fun performBulkOperation(
-        type: BulkOperationType,
-        areAllFromTheSameFolder: Boolean,
-        allSelectedFilesCount: Int?,
-        destinationFolder: File?,
-        color: String?,
-    ) {
-        // API doesn't support bulk operations for files originating from
-        // different parent folders, so we repeat the action for each file.
-        // Hence the `areAllFromTheSameFolder` set at false.
-        super.performBulkOperation(type, false, allSelectedFilesCount, destinationFolder, color)
     }
 
     companion object {
