+++ app/src/main/java/com/infomaniak/drive/ui/menu/RecentChangesFragment.kt
@@ -24,8 +24,6 @@ import androidx.fragment.app.viewModels
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import com.infomaniak.drive.R
 import com.infomaniak.drive.data.cache.FileController
-import com.infomaniak.drive.data.models.BulkOperationType
-import com.infomaniak.drive.data.models.File
 import com.infomaniak.drive.ui.fileList.multiSelect.MultiSelectActionsBottomSheetDialogArgs
 import com.infomaniak.drive.ui.fileList.multiSelect.RecentChangesMultiSelectActionsBottomSheetDialog
 import com.infomaniak.drive.utils.AccountUtils
@@ -57,18 +55,12 @@ class RecentChangesFragment : FileSubTypeListFragment() {
 
         sortButton.isGone = true
         collapsingToolbarLayout.title = getString(R.string.lastEditsTitle)
-
-        setupMultiSelectLayout()
     }
 
     private fun initParams() {
         downloadFiles = DownloadFiles()
         setNoFilesLayout = SetNoFilesLayout()
         folderId = Utils.OTHER_ROOT_ID
-    }
-
-    private fun setupMultiSelectLayout() {
-        multiSelectLayout?.selectAllButton?.isGone = true
     }
 
     override fun onMenuButtonClicked() {
@@ -82,19 +74,6 @@ class RecentChangesFragment : FileSubTypeListFragment() {
                 isAllSelected = isAllSelected
             ).toBundle()
         }.show(childFragmentManager, "ActionRecentChangesMultiSelectBottomSheetDialog")
-    }
-
-    override fun performBulkOperation(
-        type: BulkOperationType,
-        areAllFromTheSameFolder: Boolean,
-        allSelectedFilesCount: Int?,
-        destinationFolder: File?,
-        color: String?,
-    ) {
-        // API doesn't support bulk operations for files originating from
-        // different parent folders, so we repeat the action for each file.
-        // Hence the `areAllFromTheSameFolder` set at false.
-        super.performBulkOperation(type, false, allSelectedFilesCount, destinationFolder, color)
     }
 
     companion object {
